requirements.txt
firebase-admin
pandas
Flask
pytz
numpy
scikit-learn
gunicorn
google-generativeai
flask-cors